Default GPS Support Added
Added a GPS bitmask bit to indicate from flash configuration if a GPS is
installed or not. If not then the default GPS location is used from
flash memory. This isn't a 100% ideal solution but it covers most use
cases upfront except dynamically handling a GPS unit going into standby
and back. This is due to the way GPS parsing uses only the GPS UART to
trigger a parsing and thus no way to inject other info or otherwise
detect a GPS using UART (at this time).